% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPdii.R
\name{DPdii}
\alias{DPdii}
\title{Data Patcher using deletion-imputation iteration.}
\usage{
DPdii(
  data.df,
  imp = "mice",
  del_rate = 0.05,
  patch_rates = 0.1,
  elim_rates = 0.2,
  iter = 1000,
  penl = "SQD"
)
}
\arguments{
\item{data.df}{target dataset}

\item{imp}{imputation method's name ('mice' (default setting) or 'missForest')}

\item{del_rate}{deletion rate to complete dataset.}

\item{patch_rates}{patching rate. This parameter allow vector like c(0.1, 0.2,0.3)}

\item{elim_rates}{elimination rate. This parameter allow vector like c(0.1, 0.2,0.3)}

\item{iter}{the number of of iteration}

\item{penl}{calculation method of residual}
}
\value{
list of dataset

out[["patch_rate_0.1_elim_rate_0.2"]] return the dataset that its parameter is "patch_rate = 0.1" and "elim_rate = 0.2"
}
\description{
Data Patcher using deletion-imputation iteration.
}
\examples{
DPdii(iris[,-ncol(iris)], iter=1000)
out <- DPdii(iris[,-ncol(iris)],iter=1000)
out[["patch_rate_0.1_elim_rate_0.2"]]

}
